align: extract 6 large inline test blocks (rule 8 complete)
Final batch of the rule-8 sweep. Brings the inline-tests-extracted
total from 16/22 to 22/22; the rule is now fully aligned across
src/.

Files moved (large tier, 226-349 inline lines each):
- accounts/types.rs              -> types_tests.rs (348)
- proto/encoders.rs              -> encoders_tests.rs (226)
- accounts/common/encoders.rs    -> encoders_tests.rs (242)
- market_data/historical/mod.rs  -> mod_tests.rs (246)
- client/builders/sync.rs        -> sync_tests.rs (298)
- client/builders/async.rs       -> async_tests.rs (349)

Total 1,709 lines of test body relocated. Each source file now
ends with `#[cfg(test)] #[path = "..._tests.rs"] mod tests;` and
the body lives in the sibling `_tests.rs` (dedented 4 spaces from
the outer `mod tests { ... }` wrapper).

The extraction was done via a sed-based helper script
(/tmp/extract_inline_tests.sh): grep for `#[cfg(test)]`, head
everything before it, dedent the body via `sed 's/^    //'`,
append the path declaration. Mechanical 1-to-1 move; no test
behaviour changes.

Gates green on this commit: cargo fmt; cargo clippy --all-targets
under default / sync / all-features (each -D warnings).
